govwatcher-archive/requirements.txt
# Web requests and processing
requests==2.28.1
beautifulsoup4==4.11.1
lxml==4.9.1
html5lib==1.1
pyquery==2.0.0

# Web archiving tools
archivebox==0.6.2
pywb==2.7.4
warcio==1.7.4
htmldiff==0.0.6
difflib36==1.1

# Image processing
Pillow==9.4.0
opencv-python-headless==4.7.0.68

# Database access
psycopg2-binary==2.9.5
SQLAlchemy==2.0.4

# Redis client
redis==4.5.1
hiredis==2.2.2

# Web browser automation
selenium==4.8.2
webdriver-manager==3.8.5

# Concurrency and async
aiohttp==3.8.4
asyncio==3.4.3
celery==5.2.7

# Utils
tqdm==4.64.1
python-dotenv==1.0.0
pyyaml==6.0
schedule==1.1.0
click==8.1.3
jsonschema==4.17.3
validators==0.20.0

# Testing
pytest==7.2.2
pytest-asyncio==0.20.3 